[FIX] web_gantt: non determinitic pill resize year
Before this commit:
The test consist in extending a pill using drag & drop mouse event.
When the mouse drag too close from the side of the gantt view, it
will automatically scroll the view to allow to extend to further
invisible month = side scrolling

This was the case here as the browser window in test mode is set to
1366 x 768 which was making the pill very close from the side of the
window. When the mouse drop happen (after the drag) this could end
up being on a another month that the expected one due to the
automatic side-scroll.

After this commit:
We initially scroll the gantt view so that the pill is centered to
avoid side-scrolling to trigger

rb-145661

closes odoo/enterprise#82571

X-original-commit: 3d25fa00979aab4a10464abf32a2f729adfca31d
Signed-off-by: Loan Sens (lse) <lse@odoo.com>