[FIX] l10n_in_hr_payroll: use payslip id to calculate period on advice
Generating a payroll advice would use the current date in order to find
other payslips that are for the current period and use this to fill in
the period on the advice. However, if you do this for a past payslip, it
will either not find any or get the incorrect period.

Changed the behavior to use the actual payslips and get the period from
those payslips to fill in the period on the report.

opw-4582990

closes odoo/enterprise#93564

X-original-commit: 42b3bb9cb28233db6b0d4c54b87fefb2705af4a3
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Ryan Cen (ryce) <ryce@odoo.com>