[FIX] project_enterprise, timesheet_grid : Fix tests that depends on timesheet
After this commit https://github.com/odoo/enterprise/pull/74528/commits/c0af404c2b3e08ec92243f4c070f915833b4a538 some tests
are failing when we don't have timesheet_grid module installed as the
allocated hours will get recomputed so the behavior will be different.

To fix this we moved all of the tests that test allocated hours value
to timesheet_grid and adapted another test to check whether we have
the timesheet_grid module installed or not and upon this check we
change the value that we are checking.

runbot-115003

closes odoo/enterprise#88973

X-original-commit: 10f8a52f114c33129bab86b993a9d492fc6290c5
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>