[REF] *spreadsheet*: use dynamic pivot in tests
This commit changes the default insertion mode of the pivot in the tests
helpers, to use the dynamic pivot instead of the static one. It allows
to avoid to wait for the pivot to be loaded before actually inserting it.

The gain is about 20% of the time for spreadsheet tests, 15% for
spreadsheet_edition tests (enterprise) and 5% for the whole spreadsheet
test suite.

closes odoo/enterprise#81891

Task: 4655122
Related: odoo/odoo#202834
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>