[FIX] l10n_in_hr_payroll: fix traceback when generating payslip report
Currently a traceback is occurring when the user tries to
generate a payslip report.

To reproduce this issue:

1) Install `l10n_in_hr_payroll` and shift to Indian company
2) Create an employee and a bank account in the private information
3) Remove the `Bank Identifier Code` from the Bank while creating a bank record
4) Create a new payslip record with a contract running contract
4) Confirm the payslip and click the `Create Payment Report`

Error:- "TypeError: expected string or bytes-like object, got 'bool'"

As the `bank_bic` is not a required field, user can removes it. if there is
no `bank_bic`, it leads to the above traceback from the below line.

https://github.com/odoo/enterprise/blob/e2a9442ac33579c9833f65f59cf8341b3c6eafc8/l10n_in_hr_payroll/models/hr_employee.py#L24

sentry-6199042870

closes odoo/enterprise#86351

X-original-commit: cde12e0ddce6df3ea33b092b3564fbe88681b682
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>