[FIX] l10n_mx_edi: carry over serial product cancel cfdi
- Create a product with tracking by serial numbers then set a lot and serial number on the product
- Activate the option Display Lots & Serial Numbers on Invoices in the Inventory settings
- Create a sale order with the test product and confirm it
- Create the invoice and send the cfdi report
- Request a cancel and select the reason 01 Document issued with errors (with related document)
.
On the CFDI PDF report of the new invoice, there are no lines for lots and serial numbers—they are not printed.

The report uses the _get_invoiced_lot_values method from sale_stock to retrieve serial numbers. This method is first called when sending the CFDI for the parent invoice. However, when it is called for the child invoice, the method checks the quantity of the serial product, which is now depleted.

This commit introduces a hook that allows to filter the amls taken into account in get_invoiced_lot_values. In l10n_mx_edi, we do not take the amls from the parent invoice after a cfdi cancellation.

opw-4644031

closes odoo/enterprise#87875

X-original-commit: 8d7e68792857429ba7053dd6c9ccf187baddc864
Related: odoo/odoo#214608
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>