[REF] appointment: session is not a calendar props
This commit is a counterpart of odoo/odoo#202242, which cleans up
a bit the code of the calendar view and in particular removes the
"session" props of the controller. The "session" can be imported
directly, there's no need to pass it props. This commit adapts a
test, which was the only usecase where that props was given.

Followup of task-4510549

Part-of: odoo/enterprise#81657
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>