[IMP] whatsapp: add test for channel template usage
Currently when using a template on a discuss.channel two messages are
produced. The first one is the message linked to the template itself
as wa messages creates mail messages. Then a reminder log is posted
on the business document, which is the same document, which means
two messages are produced.

Adding a test to show current behavior.

TASK-3525549

Part-of: odoo/enterprise#65165
Related: odoo/odoo#196005
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>