[FIX] l10n_au_reports: Change payment state based on accountant module
Before this commit, the TestAustraliaTparReport.test_tpar
failed in a single app context with the l10n_au_reports because the "accountant"
module is not installed and the payment state will never be "in_payment".
The commit evaluate the payment state as "paid" instead if the "accountant" module
is not installed.

closes odoo/enterprise#81748

X-original-commit: f59cfef4a84bd96756a3ea6fabb7478830b53205
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>