[FIX] stock_barcode: visible source location on barcode scrap
**Problem:**
When doing a scrap from the barcode app, the source location field is not accessible.
The default source location of the scrap will be the one of the stock move, this could be a problem
if the product is located in a child location.
For instance if confirming a sale order for product A (which is stored in WH/stock/Shelf 1), the pick generated by the sale order
will be from WH/Stock to WH/Output (default pick values).
In the barcode app, if doing a scrap from this picking, the source location will be WH/stock and the client will not be able
to change that because the field does not appear.

**Steps to reproduce:**
- From the inventory app dashboard select "internal transfer"
- Create a new internal transfer
- Enter a storable product
- Set a demand quantity
- Click on "Mark as Todo"
- Open the barcode app
- Click on operations
- Select "Internal Operations"
- Select the transfer you just created
- Click on the gear icon on the top right and select "scrap"

**Current Behavior:**
The source location field is not visible

**Expected Behavior:**
The source location field should be visible and editable

**Cause of the issue:**
The source location field is not in the view
https://github.com/odoo/enterprise/blob/a9334bb551606d0151d8fc7a8f3393c3a6e312b9/stock_barcode/views/stock_scrap_views.xml#L68

**Fix:**
I added the source location field in the view

opw-4489401

closes odoo/enterprise#83707

X-original-commit: 0e741fe6559193cc2700d1623fa930f51e701843
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Pierre-Louis Hance (plha) <plha@odoo.com>