[FIX] account_intrastat: pick right decimal separator for intrastat
**Problem:**
When having products that were sold with Intrastat (which permits
the selling of products accross borders in the EU), the weight was
not correctly displayed. The weight always had the '.' decimal
separator, regardless of the language. In French, the decimal
separator is ',' but a product weighting 3 and a half kilos would
be displayed 3.5 instead of 3,5. The same can be said about the
supplementary unit column.

**Steps to reproduce:**
- Change your company to belong to the EU.
- Change the langugage to one that takes a comma as a decimal separator, like French or Spanish.
- Create a product that has a weight in the inventory tab, as well as intrastat values in the accounting tab.
- Make a sale for this product with a client in another country belonging to Intrastat and confirm the invoice.
- Go to Instrastat report in accounting in the audit reports.
- The weight is not correctly displayed.

**Cause of the issue:**
The language of the user was not taken into account when
fetching the data.

**Fix:**
The fields are now changed depending on the current language,
to match the needed decimal separator.

opw-4575332

closes odoo/enterprise#81711

Related: odoo/odoo#214208
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>