[FIX] hr_appraisal_survey: fix _compute_completed_survey_count
In dee57c5a85e26e57159a6eb66de0b2f1f581c64b, which adds prefixes to the
states of the model hr.appraisal, a search domain on an unrelated model
has unfortunately been affected too.

This commit fixes the search domain in _compute_completed_survey_count
which looks at the state of survey.user_input, which is not prefixed.

Before this commit, the Feedbacks stat button on an appraisal will
always show "0 / X", because the filtered state doesn't exist.
After this commit, it will show the correct number.

closes odoo/enterprise#81830

Task-id: none
X-original-commit: 6dc26362c78dbd144ca364a262621eb40bbfcc65
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>