[REF] web_studio: caching all relevant ir.model record data
Before this commit, we had two routes and two different calls to get some information about
the ir.model that we were editing (via the edited action).

Since we need more and more information about the ir.model (its human readable description and more), we implement
a Cache that fetches those informations with a simple [search_]read.

The information is cached because those information do not usually change in a meaningful way technically.

part of task-4153875

Part-of: odoo/enterprise#79662
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>