[FIX] sale_subscription_stock: update the qty delivered for recurring products
Issue: Validating deliveries of storable products linked to a
subscription will not update the delivered_qty untill the subscription
has been invoiced and hence wont allow you to invoice it at all if the
invoicing is based on delivered qty.

Steps to reproduce:

- Create a recurring product which inventory is tacked by quantity and
  which invoicing policy is based on delivered quantity.
- Put 10 units in stock.
- Create a monthly subscription for 1 unit of that product.
- Confirm the subscription
- Validate the associated delivery
> The qty_delivered of the sol is still at 0 and the subscription can't
  be invoiced.

Cause of the issue:

Starting from Commit b3ed4416f9b8e37604a9aa38a1f38c3504e22fdf (18.0)
we will create the first delivery linked to a subscription rather than
waiting for the cron to be triggered. These delivery moves are correctly
linked to the related sale order lines of the subscription. However,
since future deliveries will also be created and linked to that same
sol we need to consider only the outgoing and incoming moves related to
the current billing period to determine the qty_delivered. This step is
achieved with an override of the `_get_out_going_incoming_moves`:
https://github.com/odoo/odoo/blob/a6280b2a5173fad65e62af0ba72cb8721939d709/addons/sale_stock/models/sale_order_line.py#L193-L200
https://github.com/odoo/enterprise/blob/dde492541aa253771f905d0a77650aa35fea0e8c/sale_subscription_stock/models/sale_order_line.py#L27-L34
but this always leads to an empty record set since the order was never
invoiced making the `next_invoice_date` and hence the `period_end`
coincide with the start_date of the order.

opw-4628212

closes odoo/enterprise#81800

X-original-commit: 1288a6fdd9044cd444a9a2d7b9d697e52ed8c85a
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>