[FIX] project_enterprise: query missing join for related field
The query that determines overlapping tasks is making two different
Query objects: the first one for the main task and the second one for
overlapping tasks.  The second query is joined into the first one,
injecting the second one's where clause as join condition, but that
composition does not work if the second query has extra joins.  Those
extra joins come from related fields that may occur from the domain.

The fix consists in recomposing the query with a join condition that
does not depend on a dynamic domain, and injecting the domain in the
where clause of the main query.  With this strategy, the joins that
appear because of the related fields in the domain are part of the main
query, and are no longer lost by the composition.

Note also that the test
test_same_user_overlap_with_allocated_hours_less_than_workable_hours was
actually broken, because the field allocated_hours wasn't flushed when
doing the SQL query.  The new query flushes the field, and the test has
been fixed.

closes odoo/enterprise#85894

Related: odoo/odoo#210601
Signed-off-by: Raphael Collet <rco@odoo.com>