[IMP] documents: show warning when sharing to external user
This commit adds a warning when one tries to share a document to some external
users but the `access_via_link` parameter is set to `none` which will prevent
the users to actually access the document. It also adds some visuals to quickly
target the problematic users.

Task-4480388

Part-of: odoo/enterprise#84750
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>