[FIX] pos*: skip preview screen not doing everything
pos*: l10n_it_pos, whatsapp_pos

When setting the options automatic receipt printing with skip preview
screen in a pos restaurant, after the payment, the user is redirected
to the default screen instead of the floor screen.
Currently, some actions are not done during the skipping of the receipt
screen.

This commit make sure the actions are done when the skip preview screen
is enabled.

closes odoo/enterprise#81768

Task-id: 4170340
Related: odoo/odoo#202548
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>