[FIX] l10n_de_pos_cert: fix empty receipt
When automatically sending the receipt in a DE PoS, there was 2 receipt
sent to the printer, one of them was empty and showing an error.

Steps to reproduce:
-------------------
* Setup a fiskaly PoS
* Activate the automatic receipt printing
* Open PoS and make an order
* Pay the order
> Observation: Two receipt are sent to the printer, one of them is empty
and showing an error.

Why the fix:
------------
We were calling the `super` method twice. This was causing the receipt
to be printed twice. We now call the `super` method only once.

opw-4520201

closes odoo/enterprise#82248

X-original-commit: 01609b3a65cae6537787a45d7a939bd79682a5dc
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>