[IMP] l10n_ar_edi: improve connection test popup window and style
The previous connection test window showed up as an UserError with
the title "Invalid Operation" which can be confusing to users when
all services show as available.

Replace the UserError with a RedirectWarning that has a better title
message ("Odoo Warning"). As it needs to redirect to somewhere, we
redirect back to the current accounting page in settings.

task-4288536

closes odoo/enterprise#80415

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>