[FIX] mrp_workorder: run procurements once when adding component in shopfloor
Step to reproduce:

- In the settings enable Multi-step routes
- Inventory > Configuration > Warehouse Management > Warehouses
- Put you warehouse in manufacturing in 2 steps
- Create a bill of material for a final product FP with one raw:
    - 3 x COMP (storable product with 10 units in stock)
- Create and confirm an MO for 1 unit of FP.
> This create a picking from stock to pre-prod for 3 unit of COMP
- In the shopfloor on your MO click on the 3 dots and add components
- Register a new unit of COMP
> The picking from stock to pre-prod was updated twice, hence for 5 units

Cause of the issue:

When you add a product from the shopfloor, we create a new move
from pre-prod to virtual/production with the corresponding
`product_uom_qty` of 1 and confirm it:
https://github.com/odoo/enterprise/blob/fc1fb4c56da916b165c43d8f6b6a4b903733a12f/mrp_workorder/wizard/additional_product.py#L63-L68
Since the procure method of this move has been adjusted, this
confirmation will correctly create and run a procurement to generate a
move from stock to pre-prod. Note that this part of the flow is strictly
necessary if the additional product is not already part of the component
raw of the MO. However, since in the present case  there is already move
raw from pre-prod to virtual/production associated to that product,
the action confirm will also merge our additional move with the current
existing one and hence modify its `product_uom_qty`. However, an
override of mrp ensures that when such modification happen, we
should also run the procurement to ensure that modifying the demand of a
move will also update the related pickings:
https://github.com/odoo/odoo/blob/7def831bea18a91e4fa0f9c6aa6de34f4d6d18c8/addons/mrp/models/stock_move.py#L413-L419
Since we already run that same procurement we should bypass this call in
our case.

opw-4562965

closes odoo/enterprise#84296

X-original-commit: 5de2795ad4cc7af4cd76972e6e087f5118d0cd34
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>