[FIX] stock_barcode_product_expiry: display expiration date
When scanning a package, its content is added on the view, but we
don't display the expiration date.

To reproduce the issue:
1. In Settings, enable:
   - Packages
   - Expiration dates
2. Create a product
   - Tracked by lot
   - Expiration date
3. Update its quantity
   - Lot L
   - Package P
   - Quantity > 0
4. Set an expiration date on L
5. Open a new delivery through the barcode app
6. Scan P

"Error": a line is added for the product, with its lot L, but we
don't display the expiration date. Once the line is saved (for
instance, through the edit button), we will finally display it.

This is because we don't provide any information about the
expiration date when converting the package quant into relevant data.

OPW-4516410

closes odoo/enterprise#82570

X-original-commit: 605d379c8fa630620e2ee859e0c0b133b1e34100
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>