[FIX] account_accountant_batch_payment: test reset of batch payment
Test that resetting a statement line that was merged with a batch
payment also resets the payments, especially for payments without
an outstanding account.
It wasn't the case because there is no real link between the batch
payment and the statement line, and neither between the statement line
and the payment itself.

The fix is in community.

closes odoo/enterprise#82258

X-original-commit: d3cb3c268bbdc1ebcd27c59e4d1ca1fcf3682af9
Related: odoo/odoo#203582
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>