[FIX] sign: test_sign_flow_tour
In this commit, we fix the latest step of tour. We can't click on
primary or secondary button ... make no sense.
As the primary button download the pdf file and there is no test to
ensure the file is well downloaded, it make more sense to click on
close the modal.
With the fix in https://github.com/odoo/odoo/pull/202089 and without
the correction, the tour was timeout exceeded.

closes odoo/enterprise#81831

X-original-commit: 04975eb8adfcd386efc87f1df0a05cabe57075e7
Related: odoo/odoo#202680
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>