[REF] sale_*: do not autosubscribe customer to SO
Update to SO access rights update, as they are not based on followers
anymore for portal flows. Commercial entity of the customer is now
the base rule.

Also remove auto subscription of customers, as done in community code
of sale.

Task-4655022

Part-of: odoo/enterprise#82683
Related: odoo/odoo#204326
Related: odoo/upgrade#7479
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>