[REM] account_asset: first_depreciation column in depreciation schedule
The "First Depreciation Date" column is misleading when importing assets from previous accounting systems, as it often displays an incorrect date.

As if the asset is created in Odoo, the date usually matches the acquisition date. But if we import assets that were already depreciated in another system, the date is often wrong.

This change removes the "First Depreciation Date" column to prevent confusion and ensure more accurate reporting for both Odoo-created and imported assets.

task-4657876

closes odoo/enterprise#81906

Signed-off-by: William André (wan) <wan@odoo.com>