[IMP] project_account_budget: show clickable buttons in project dashboard
Before this commit, the budget sections displayed in the project right
side panel are not displayed as clickable element, which means the
cursor does not indicate the element is clickable.

This commit adds a `href` on those buttons to make sure the cursor shows
those elements are clickable as expected.

task-4593181

closes odoo/enterprise#81986

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>