[IMP] industry_fsm: reduce font size of customer preview stat button
Before this commit, the customer preview stat button displayed in fsm
task form view has a font size bigger than the other stat buttons and so
the text of that button could be cropped compared of the other buttons.

This commit reduces the font size to make the same font size than the
other buttons.

task-4662072

closes odoo/enterprise#81988

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>