[IMP] helpdesk: adapt text of 'log note' onboarding tour step
This commit adapts a step asking to click on `log note` button to send
an internal note. We no longer mention the user can select a channel
because that feature does not ping the users inside the channel
selected, it is just to have a button to redirect the user to that
channel.

task-4662201

closes odoo/enterprise#81993

Related: odoo/odoo#203081
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>