[FIX] website_knowledge: reenable (un)folding articles public sidebar
This commit fixes an issue with the public sidebar. When you unfold an article
the unfolding method was never linked to all the elements needed to unfold any
children article.

Now when clicking on the unfold button, we are adding the event on each children
of the unfolded articles.

task-4668046

closes odoo/enterprise#83044

X-original-commit: 5facb11652dbdbd82db349015cc7eebdcc5c4a9f
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>