[FIX] sale_commission*: avoid applying twice the currency rate
Before this commit, the commission applied the currency rate to convert
the commission to the currency of the current company. As the achivement
was already converted to the current company, it was not neccesary to do
it twice. Most of the time, that final currency rate was equal to one
because the company of the achievement was already equal to the company
of the commission. Therefore the rate was 1.

Taskid: 4652771
X-original-commit: 70293a8cdc30d5b3f6a532c87b7936697b4e16a2
Part-of: odoo/enterprise#82508
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>