[FIX] l10n_pe_edi: fix PriceTypeCode with more decimal digits for price
Steps to reproduce:
- Install Accounting and l10n_pe_edi
- Switch to a Peruvian company (e.g. PE Company)
- Go to "Settings / Technical / Database Structure / Decimal Accuracy"
- Set 4 digits for "Product Price"
- Create a product with a lower price than 0.01 (e.g. 0.0045)
- Create an invoice:
  * Customer: [a Peruvian customer] (e.g. Comercial Constructora los Patitos S.A.)
  * Product: [the created product]
  * Quantity: 100
- Confirm the invoice
- Process the invoice by the E-invoicing service

Issue:
An error is triggered because "PriceTypeCode" is set to "02", which is
the code used to declare that it is free.
But in this case, it is not free.

Cause:
"PriceTypeCode" is computed by using "is_zero" method of the currency
on the unit price.
As the decimal accuracy of the currency is set to 2 and the unit price
is 0.0045, "is_zero" is returning True.

Solution:
Compute "PriceTypeCode" from the line subtotal instead of the unit price.
This is not perfect as it can also be lower than 0.01, but it's less
probable.

opw-4608671

closes odoo/enterprise#83463

X-original-commit: c89c52b980d851dffd06dabfba6bd9c662e0659b
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>