[FIX] l10n_nl_reports_sbr{_icp}: keep previous XBRL version
[This commit] updated the XBRL version of the Dutch tax reports from the
2024 version (NT18) to the 2025 version (NT19). However, users still
need to be able to submit reports for 2024 using the previous scheme.

In order to allow this, we keep both the 2024 and 2025 version of the
XBRL template and choose the right one depending on the year of the
report. In subsequent years, we can then remove the oldest version and
add the newest version.

The report IDs are suffixes with the year.

opw-4600111
opw-4664515

[This commit]: https://github.com/odoo/enterprise/commit/f1710461f37f3560486a1410f9b9ff420b052b7b

closes odoo/enterprise#82166

X-original-commit: d891b57460ee57fd6326e31439d62c81ae5d87f2
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>