[FIX] delivery_iot: send seperate printing jobs when validating deliveries
Currently, when validating deliveries, if multiples attachments are posted
in the same chatter message, the printing jobs do not complete.

Steps to reproduce:
-------------------
* Install fedex
* In operation types, select Delivery Orders then hardware
* In print on validation, enable Carrier Labels and Export Documents
* In the shipping methods, select fedex international
* Change Label format to ZPL11
* Connect the database to an iot box
* Set up the shipping labels to be send on zebra printer
* Set up commercial invoice (/invoice) to be sent to another printer
* Create a sale order, using fedex international as shipping
* Confirm the SO
* Select the delivery
* Validate the delivery
> Observation: Nothing prints, jobs are sent to CUPS bu not printing

Why the fix:
------------
https://github.com/odoo/enterprise/blob/8075101192fb81a78f2a984cf2adf67fc77c0194/delivery_fedex_rest/models/delivery_fedex.py#L192-L196

As show, when sending the shipping, if the invoice is generated, it gets added
to the attachements. The function `message_post` will post all attachments in
the same chatter message.

In `delivery_iot` the `message_post` function is overriden to allow sending the
printing jobs to the iot.

https://github.com/odoo/enterprise/blob/8075101192fb81a78f2a984cf2adf67fc77c0194/delivery_iot/models/stock_picking.py#L33-L46

Here, `attachments_names` will both contain `Label` and `ShippingDoc`.
Since we enter the first if condiftion, the report is related to the
shipping labels. Then we send the data related to all attachments to
the device linked to the shipping labels, aka the zebra printer. Since
the other attachment is of type PDF, the zebraprinter does not know what
to do with it and ends up printing nothing.

Since we can have multiple different attachment, the if/elseif condition
does not make sense. By breaking the attachments, we can send both reports
seperately, while still keeping all attachments in the same chatter
message.

opw-4546715

closes odoo/enterprise#84077

X-original-commit: b39a6c7c46ed16cd1fd01d498ecc1bfe6a91bd01
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>