[FIX] *inter_company_rules: sync discount in SOL/POL
Previously, POL did not have a `discount` field which meant that the
`discount` of a SOL was transferred in the `price_unit` of the POL.

For example, the following SO:

Product     | Unit Price | Discount | Tax Excl.
------------|------------|----------|----------
Product A   | $1000.0    | 1O.0%    | $900.0

Would be converted into the following PO

Product     | Unit Price | Tax Excl.
------------|------------|----------
Product A   | $900.0     | $900.0

However, as of V17, it is possible to add discount on POL, but are
not taken into account when converting into SOL. With this commit, the
discounts are now synchronized when going from SOL->POL and inversely.

opw-4446288

closes odoo/enterprise#83193

X-original-commit: 2b9b4654a939a47cc4a877c68dca48acee28f19d
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>
Signed-off-by: Lionel Piraux (lipi) <lipi@odoo.com>