[REF] timesheet_grid: review timer header component
Before this commit, multiple bug fixes have been made in the timer
header when this component has been mounted in the list view since the
list model conflicts with the one used in timer header component.

This commit completely reviews the timer header component. It also
introduces a new service called timesheet timer service which is the
timer service but with additional data needed for the timer header
component.

task-4284216

Part-of: odoo/enterprise#81686
Related: odoo/odoo#202313
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>