[FIX] account_reports_cash_basis: prevent test in invoicing enterprise
The modified test fails when only the Invoicing Enterprise module is
installed, without the Accounting module. This failure occurs because
the outstanding account's code is already in use by another account.
This test is irrelevant when only Invoicing Enterprise is installed,
as the general ledger is accessible only with the Accounting module.
Therefore, we skip this test if the Accounting module is not installed.

closes odoo/enterprise#82122

Runbot-error: 134665
X-original-commit: 0418b5f38acae453eef14472e9ef057636ed8ec0
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>