[FIX] sign: check model for sign item automatic value
Before this commit, when a sign field type was attached to a model,
it could be automatically completed if the sign.request was linked to
another model.

Steps to reproduce:
1) Create a sign field type: text, model: sale.order, field: reference
2) Add this field in a sign.template.
3) Open a manufacturing order (mrp.production), create a sign request
   for it
4) sign the MO, the reference field is automatically filled.
The method getting the value to fill, don't check the model. As a reuslt
if the same field exists, it will be filled.

closes odoo/enterprise#84348

Taskid: 4672517
X-original-commit: 96d1361013c719359c2c97a38e31fec5fbc045ca
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>