[FIX] helpdesk: auto-assign tickets from website form
**Steps to reproduce**
On runbot:
1. Activate the website form for the "IT Support (Auto Assignment
by Tags)" team.
2. Add the "Tags" field to the form, allowing visitors to select
tags for a ticket.
3. Submit a ticket with the "Repair" tag selected.
Actual: the ticket is not auto-assigned.
Expected: Mitchell Admin is auto-assigned, the same way as if the
ticket was created from the backend.

**Cause**
When using the website form, a `SET` command is used to set the
tags, while we are currently only expecting a `LINK` command.

opw-4646654

closes odoo/enterprise#83403

X-original-commit: f9279d14920b41c1372ae00ac53690db10f99b6c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>